整理代码，去掉重复的EditorMaterialChunk
EditorMaterialChunk和MaterialBlockChunk都是0xAFFF。
