+++ Functions/CSharpGL.FileParser._3DSParser/Chunks/MaterialBlockChunk.cs
@@ -8,5 +8,33 @@ namespace CSharpGL.FileParser._3DSParser.Chunks
 {
     public class MaterialBlockChunk : ChunkBase
     {
+        //public string MaterialName;
+        //public ThreeDSMaterial Material;
+        //public ColorChunk AmbientColorChunk;
+        //public ColorChunk DiffuseColorChunk;
+        //public ColorChunk SpecularColorChunk;
+        //public PercentageChunk ShininessChunk;
+
+        public override void Process(ParsingContext context)
+        {
+            ProcessMaterialChunk(context);
+        }
+
+        void ProcessMaterialChunk(ParsingContext context)
+        {
+            var reader = context.reader;
+            var chunk = this;
+
+            while (chunk.BytesRead < chunk.Length)
+            {
+                ChunkBase child = reader.ReadChunk();
+                child.Parent = chunk;
+                this.Childern.Add(child);
+
+                child.Process(context);
+
+                chunk.BytesRead += child.BytesRead;
+            }
+        }
     }
 }
